(REFACTOR) user/seatsearch MVI 통일, 컴포넌트 분리 및 UseCase 추출 (#20)
